+++ lib/src/common/java/testutil/NamedCsvRecordAssert.java
@@ -1,8 +1,10 @@
 package testutil;
 
 import org.assertj.core.api.AbstractAssert;
+import org.assertj.core.api.Assertions;
 import org.assertj.core.api.InstanceOfAssertFactory;
 import org.assertj.core.api.MapAssert;
+import org.assertj.core.api.OptionalAssert;
 import org.assertj.core.api.StringAssert;
 
 import de.siegmar.fastcsv.reader.NamedCsvRecord;
@@ -38,6 +40,11 @@ public class NamedCsvRecordAssert extends AbstractAssert<NamedCsvRecordAssert, N
     public StringAssert field(final String name) {
         isNotNull();
         return new StringAssert(actual.getField(name));
+    }
+
+    public OptionalAssert<String> findField(final String name) {
+        isNotNull();
+        return Assertions.assertThat(actual.findField(name));
     }
 
 }