+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for different field count for NamedCsvReader
 - Metadata for OSGi capability
 - Support for optional BOM header when reading CSV files
+- Method `NamedCsvRecord.findField` for optional field access
 
 ### Changed
 - Updated from Java 8 to Java 11